changing from last years JSON and changing it to match the mesh's current code
This should match the current mesh code we have, I added another sensor reading called soilM so that we can have a separate reading for soil only.

+++ SoilSendFile/readsoil.h
@@ -1,6 +1,6 @@
 #ifndef READSOIL_H
 #define READSOIL_H
 
-void read_soil_rh(int pin, int & rh);
+void read_soil_rh(int pin, float & rh);
 
 #endif
